# docs/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MCI Dashboard REST API",
    "description": "REST API for MCI Banking Dashboard - Real-time metrics for Financing, Saving, and Deposito",
    "version": "1.0.0",
    "contact": {
      "name": "MCI Development Team",
      "email": "dev@mci.co.id"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8089/api",
      "description": "Local Development"
    },
    {
      "url": "https://dashboard.mci.co.id/api",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/v1/dashboard/chart/{type}": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get Chart Data",
        "description": "Retrieve historical chart data for financing, saving, or deposito",
        "operationId": "getChartData",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "description": "Chart type: financing | saving | deposito",
            "schema": {
              "type": "string",
              "enum": ["financing", "saving", "deposito"]
            },
            "example": "financing"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved chart data",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "labels": ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
                    "values": [140000000000, 142000000000, 145000000000, 143000000000, 146000000000, 148000000000, 150000000000, 149000000000, 151000000000, 150000000000, 152000000000, 150000000000],
                    "noa": [1400, 1420, 1450, 1435, 1460, 1480, 1500, 1495, 1510, 1505, 1520, 1500],
                    "growth": [0, 1.43, 2.11, -1.38, 2.10, 1.37, 1.35, -0.33, 1.34, -0.66, 1.33, -1.32]
                  },
                  "meta": {
                    "type": "financing",
                    "generated_at": "2026-04-24T10:30:00+07:00",
                    "cache_ttl": 300,
                    "version": "v1"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid chart type",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid chart type. Valid: financing, saving, deposito"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/dashboard/clear-cache": {
      "post": {
        "tags": ["Dashboard"],
        "summary": "Clear Dashboard Cache",
        "description": "Clear all cached dashboard data (admin only)",
        "operationId": "clearCache",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cache cleared successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dashboard cache cleared successfully",
                  "meta": {
                    "generated_at": "2026-04-24T10:30:00+07:00",
                    "version": "v1"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Unauthorized - Admin access required"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication for protected endpoints"
      }
    }
  }
}